Move controller logic from resource to controller
